scripts/generating_topic_summaries/requirements.txt
google-cloud-storage==2.19.0
functions-framework==3.8.2
vertexai==1.71.1
google-cloud-aiplatform==1.71.1